run_methyltfr

This function is a wrapper function to calculate the deviation in transcription factor footprint base for all given motifs using parallel package

run_methyltfr(
  sample_ann,
  sample_dir,
  genome = "hg38",
  tf_bindsites = NULL,
  gcfreqs = NULL,
  gc_dist = NULL,
  sampleColName = "bedFile",
  chunkSize = 20,
  full_path = FALSE,
  annfile = NULL,
  threads = 1,
  enhancer = NULL,
  filetype = c("EPP", "BisSNP")
)

Arguments

sample_ann

- a tab seperated file contains sample annotations

sample_dir

- directory where all bed file and annotation file stored

genome

- human genome version default: hg38

tf_bindsites

- a GenomicRange object contains tf binding sites positions from (methylTFRann)

gcfreqs

- GC bin frequency tables (matrices for multiple motif) from (methylTFRann)

gc_dist

- Genome wide GC distribution from (methylTFRann)

sampleColName

- column name of the sample bed file in the annotation file

chunkSize

- chunk size for parallel processing of motifs

full_path

- if TRUE, the bed file path in the annotation file is full path

annfile

- if provided, the sample annotation file is not read from the sample_dir

threads

- thread count for parallel processing

enhancer

- Specific regions like distal motif

Value

deviation score matrix for all samples